reduce teset boilerplate two
